# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateAccountRequest:
    properties:
      account_number:
        type: string
      account_type:
        type: string
      color:
        type: string
      currency:
        type: string
      name:
        type: string
      parent_account_id:
        type: string
    type: object
...
  dto.UpsertShareRequest:
    properties:
      login:
        type: string
      permission:
        type: string
    type: object
  entity.Account:
    properties:
      account_number:
        type: string
      account_type:
        type: string
      balance:
        description: Joined
        type: string
      client_id:
        type: string
      closed_at:
        type: string
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      investment_account_id:
        description: Joined
        type: string
      name:
        type: string
      parent_account_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  entity.AccountPatch:
    properties:
      color:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  response.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/response.ErrorBody'
    type: object
host: api.goen.wisteria.local
info:
  contact: {}
  description: REST API for Goen Project
  title: Goen API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Retrieve a list of accounts for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Account'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Create a new banking or manual account for the user
      parameters:
      - description: Account Creation Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Account
      tags:
      - Accounts
  /accounts/{accountId}:
    delete:
      description: Delete an account (and its associated dependencies according to
        business rules)
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete Account
      tags:
      - Accounts
    get:
      description: Retrieve specific account properties by ID
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Account'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Account
      tags:
      - Accounts
    patch:
      consumes:
      - application/json
      description: Partially update account information
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Account Patch Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.AccountPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update Account
      tags:
      - Accounts
  /accounts/{accountId}/shares:
    get:
      description: List active sharing links provided to other users for this account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Account Shares
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: Add or update view/admin permissions for another user on an account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Upsert Share Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Upsert Account Share
      tags:
      - Accounts
  /accounts/{accountId}/shares/{userId}:
    delete:
      description: Remove access permissions given to another user for this account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Target User ID or Username
        in: path
        name: userId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Revoke Account Share
      tags:
      - Accounts
  /accounts/balances:
    get:
      description: Get current financial balance calculations for each account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Account Balances
      tags:
      - Accounts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"